ninja: Fix build_by_default=false targets in tests
Previously, build_by_default=false targets would not be built at all
even if they were used in a test (as the exe or as a command-line
argument) which would lead to a test failure.

Now we look into all the defined tests and add all CustomTargets and
BuildTargets used in them to the list of build_by_default targets.