Change the order of link_args when cross_compiling with ninja.
A user may want to add libraries to link with in the (c|cpp)_link_args
property of the cross-compile file.
Those libraries should be at the end of the command line due to reference
resolution mechanism of the compiler.

By moving the cross_args after LINK_ARGS we are sure that specific
cross-compilation libraries are at the end of the command line.

See [github PR #1363](https://github.com/mesonbuild/meson/pull/1363) to
have the context of this change.