tests/common/129: Fix configure_file invocation
There is no way for us to know that 'source.c' is a file in the source
tree if it's a string. It needs to be a file object.

This used to work earlier because we used to incorrectly run the
configure_file() command in the source dir (!) instead of the build
dir. This had nasty side-effects such as creating files in the source
tree unless you specified an absolute path...