tests: Use #!/usr/bin/env python3 for all scripts
We automatically convert that to use sys.executable now which is
always available on all platforms (because we're running with it).

On some platforms like NetBSD, `python` doesn't exist, and you must
use a specific python version. On most other distros, `python` is
Python 2, and we don't want to depend on that.

Closes https://github.com/mesonbuild/meson/issues/695

All these scripts were being used as `find_program()`, so we do not
lose any test coverage by doing this.