Add manual-linking tests to Linux and Windows
In this test, we try to manually link against the generated library to
create an executable and then run it to verify that it works.

Also test for all possible library versioning in the versioning tests on
Windows. Even though they yield the same dll naming, we should still
test it.