do not use universal_newlines for subprocess.Popen() for Python < 3.6
universal_newlines prior to Python 3.6 uses the default encoding
returned by locale.getpreferredlocale(False), which might not be the
encoding that the console actually uses.
Using the wrong encoding leads to an Exception, which can be seen on
a Windows machine that is set to German and uses an utf-8 console when
running the meson test suite (and possibly at other places).