Ignore install as non-existent user or group
'test cases/common/12 data' and 'test cases/common/66 install subdir' both
try to install something as 'root'.  This user doesn't exist on Cygwin.

Perhaps we should automatically convert this to some other user, but it's
not clear which. For real projects, it's possibly better for the meson.build
to explicitly handle this...

'test cases/common/12 data' also tries to install something with GID 0.
There's no guarantee this group exists.