Allow more convenient values for UserBooleans
Allow the user-specified boolean options to use more commonly used
values including yes/no, on/off, 1/0. This is based on a similar feature
in CMake which turned out quite convenient for users. Especially, this
makes it possible for Meson to integrate better in existing build system
logic in Gentoo without having to reinvent the wheel for yet another
implementation of booleans (Gentoo uses yes/no for some reason).