Add a new method for arrays: array.found()
This allows users to check if all the dependencies in a list of
dependencies were found. Also allows you to use [] as a placeholder
for a dependency that was not found. We already recommend the
following idiom:

    feature_deps = []
    if get_option('build-feature') != 'no'
        feature_deps = [dependency('dep1', required : get_option('build-feature') == 'yes'),
                        dependency('dep2', required : get_option('build-feature') == 'yes')]
    endif

    ...

    executable('feature', 'src.c', dependencies : feature_deps)

This would allow people to do this:

    if feature_deps.found()
        executable('feature', 'src.c', dependencies : feature_deps)
    endif

Instead of having to check the entire list of deps. This is also
useful when you want to only check a single dependency since it avoids
having to store a boolean variable called 'found_dep1' or similar for
every feature that you check for.

As a bonus, this can also be used for checking if a list of external
programs were all found.