mesonconf: reorder output columns
mesonconf prints build dir information in the order of

  'Option' 'Description' 'Current Value' and, optionally, 'Possible Value'.

The Description tends to be longer and push the following values far
right, sometimes way far than fits in the full screen size of a
terminal on a FullHD monitor.

Experienced users know which options they want to change without
looking at the description string however they need to check the
current values for sure.

This patch moves the description to the last column.  Now mesonconf
prints options in the following order:

  'Option' 'Current Value' 'Possible Value' 'Description'

To implement this, mainly print_aligned() is modified.  The second
argument of the function is changed from

 - array of array of string to
 - array of dict with 'name', 'descr', 'value', and 'choices,

which maps to 'option', 'description', 'current value' and 'possible
values', respectively.

Since the position of the possible values are moved before its
description, the presence and the length of it affects header as well
as the following description.  Thus, we cal curate it before printing
the header.

To avoid re-calculation, we keep string version of the values and
flattened version of the possible values _in the given array_, which
means that now the print_aligned() function modify the the given
array.  The current callers do not use the passing array.  So there
should be no bad effects.