Don't force target names to be unique everywhere
The only reasons this was added were:

1. Building with a flat layout is impossible
2. Installation into the same destdir would overwrite the previous file

We now error out on duplicate target names only when the layout is
flat, and we error out on install if we try to install two files of the
same filename into the same directory.

This is actually better than the previous behaviour because now we
explicitly ensure that we never install any two identically-named files
into the same directory.