Add a helper for maintaining backwards-compat code
We now check the meson_version specified in project() (per subproject),
and if it is older than the version in which the compatibility-breaking
change was made, we use the older behaviour and print a warning.

Hence, if you set `meson_version : '>=0.41.0'` in project(), you will
not be allowed to use path separators in target names.

This allows us to have a consistent warning message for deprecations,
allows us to track deprecations across releases, and will also make it
easy to remove the fallback code when we decide to stop supporting
backwards-compat with a specific old version of Meson.

Will also be useful for https://github.com/mesonbuild/meson/pull/761