RPM macros: export LANG=c.UTF-8
meson is spitting out warnings if the build system does not use a
UTF-8 locale. As users of the macros are most likely to be in some
kind of headless build systems, locale is not of a big importance per
se, so we can simply export LANG=C.UTF-8 and silence meson's warnings.