appveyor: Don't skip doc-only builds
This was a good idea in theory, but in practice it has made Codecov
completely useless because doc-only commits would reduce the coverage
calculated by Codecov and PRs would randomly have messages saying that
the coverage has gone up or down by a random percentage simply because
the windows builds weren't run in the last commit.

Till Codecov also has an option to skip doc-only commits, we cannot
make this change. To skip all CI, you can put "[skip ci]" or
"[ci skip]" in the commit message. This will also work with Codecov.