dist: use git ls-files
This avoids temporary repository clone, and submodule update etc, it
should work fine offline.

It allows to test modified files too, which was the primiary reason
why I started this. However, to keep current behaviour of meson, that
only dist the files committed, we may want to use a different target,
or a configure/setting variable?

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>