Add option to search for .pc files in $prefix/$libdir/pkgconfig
pgk-config .pc files are written by meson to
$prefix/$libdir/pkgconfig.

However, that same path is not used by meson to look for dependencies.

When multiple projects are installed to the same custom prefix,
meson knows how to write .pc files to
$prefix/$libdir/pkgconfig in all projects, but meson is not able
to find already installed .pc files in that prefix.

The use of subprojects and wraps so missing dependencies are installed
to the $prefix is not practical, as the dependency() will not be
detected and it will always be reported as missing (pkg-config does not
look into $prefix/$libdir/pkgconfig).

As a workaround, the user building the projects needs to be smart
enough to know where meson is installing the .pc files and
define in advance PKG_CONFIG_PATH=$prefix/$libdir/pkgconfig,
so .pc files can be found.

Given that meson *decides* where to write the .pc files inside
the $prefix, it is reasonable that meson checks *that same path*
when looking for dependencies.

Telling the user to find the right PKG_CONFIG_PATH value
is counterintuitive, as if meson ever decided to change the
installation path it would force all users to fix their extra
environment variables.

pkg-config now gives a special treatment to the /usr/local and
/usr prefixes over the rest, as they are always searched. pkg-config
is not aware of existing $prefix, unless the user (or meson) says
so.

meson currently gives a special treatment to the $prefix when writing
.pc files, but does not give the equivalent treatment to $prefix when
looking for .pc files. This commit addresses that, by allowing the
definition of a `search_prefix` argument to `dependency()`, so
the meson instruction:

dep = dependency('core_project', search_prefix: true)

looks for core_project.pc in the PKG_CONFIG_PATH (if defined by the user),
in $prefix/$libdir/pkgconfig and in PKG_CONFIG_LIBS (in that order).

By default `search_prefix` is false, so backwards compatibility is preserved.