i18n: install .mo files for immediate use in the builddir
Previously, we just generated the $lang.gmo files and then just copied those
into $destdir/$lang/LC_MESSAGES/$pkgname.mo. But to be immediately useable,
they need to be in that layout inside the builddir, so translations are
testable without having to install them.

This patch merely adds the generation of those .mo files, they are not
installed or otherwise used.