Implement use keyword for compiler.has_header method
Sometimes we want to check that a header can be used in compilation, not
only that it exists. Add a new boolean keyword 'use' for has_header
method of compiler class. When this keyword is true, we try to compile
the header, when false we only check that we can preprocess it. Create
method check_header to check for compilation of header.