Add a small test case for compiler.has_header use keyword
In this case we have two headers: one can be used in compilation and the
other will cause an error if compiled. We then test that the non-usable
header will make compiler.has_header with the keyword use: true return false.
Finally, we test that the usable header will make compiler.has_header
with keyword use: true return true.