Fix quoting of cflags/ldflags for gtkdoc-scangobj
The gtkdoc-scangobj program takes a --cflags and --ldflags
arg each of which takes a list of arguments which will then
be passed to cc/ld. If any of these args include paths we
must be sure to escape then e.g. the current code execs

  gtkdoc-scangobj '--ldflags=-L/some/path with/spaces -lfoo'

which causes exec of

   ld '-L/some/path' 'with/spaces' '-lfoo'

What we need instead is

  gtkdoc-scangobj '--ldflags="-L/some/path with/spaces" "-lfoo"'

which causes exec of

  ld '-L/some/path with/spaces' '-lfoo'

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>