options: Add a permitted_kwargs function
I've typo'd "value" for the last time, options needs a kwargs validator.
This validator is slightly different than the one used by the main
parser, since it operates on a much simpler representation than the
other one does, and they are not interchangeable.

This also changes the optinterpreter to use pop on 'type' and
'description' so that they're not passed to the validator as kwargs.